+++ CHANGELOG.md
@@ -34,6 +34,7 @@ adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 - Fix linker errors for build target `x86_64-linux-unknown-gnu` by updating `open62541-sys` to
   version 0.5.4 ([#288](https://github.com/HMIProject/open62541/issues/288)).
+- Improve handling of and recovery from connection loss in `AsyncClient`.
 
 ## [0.10.1] - 2025-10-29